# Src/Web/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Trenini.net API",
    "description": "A web API for model railways collection",
    "termsOfService": "None",
    "contact": {
      "name": "Carlo Micieli",
      "url": "https://carlomicieli.github.io",
      "email": ""
    },
    "license": {
      "name": "Use under LICX",
      "url": "https://example.com/license"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/CatalogItems/{slug}/rollingStocks/{rollingStockId}": {
      "delete": {
        "tags": [
          "CatalogItems"
        ],
        "operationId": "CatalogItems_DeleteRollingStock",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "rollingStockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      },
      "put": {
        "tags": [
          "CatalogItems"
        ],
        "operationId": "CatalogItems_EditRollingStock",
        "parameters": [
          {
            "name": "slug",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "nullable": true
            },
            "x-position": 1
          },
          {
            "name": "rollingStockId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "guid"
            },
            "x-position": 2
          }
        ],
        "requestBody": {
          "x-name": "request",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditRollingStockRequest"
              }
            }
          },
          "required": true,
          "x-position": 3
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditRollingStockOutput"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "422": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "extensions": {
            "type": "object",
            "nullable": true,
            "additionalProperties": {}
          }
        }
      },
...
      "Slug": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "value": {
            "type": "string"
          }
        }
      },
...
      "EditRollingStockOutput": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "slug": {
            "$ref": "#/components/schemas/Slug"
          }
        }
      },
      "EditRollingStockRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "epoch": {
            "type": "string",
            "nullable": true
          },
          "category": {
            "type": "string",
            "nullable": true
          },
          "railway": {
            "type": "string",
            "nullable": true
          },
          "className": {
            "type": "string",
            "nullable": true
          },
          "typeName": {
            "type": "string",
            "nullable": true
          },
          "roadNumber": {
            "type": "string",
            "nullable": true
          },
          "passengerCarType": {
            "type": "string",
            "nullable": true
          },
          "serviceLevel": {
            "type": "string",
            "nullable": true
          },
          "lengthOverBuffer": {
            "nullable": true,
            "oneOf": [
              {
                "$ref": "#/components/schemas/LengthOverBufferRequest"
              }
            ]
          },
          "dccInterface": {
            "type": "string",
            "nullable": true
          },
          "control": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "LengthOverBufferRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "millimeters": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "inches": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          }
        }
      },
...
    },
    "securitySchemes": {
      "JWT Token": {
        "type": "apiKey",
        "description": "Copy 'Bearer ' + valid JWT token into field",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}